Fix IsotopePicker fields
In previous commit was forgotten to add a property in a wrapper in order
to get Picker clipped to its edges.

This commit add the property used in the source of "trick", but
`backgroundColor: 'transparent'` also make it works.

See original idea at https://github.com/facebook/react-native/issues/7817#issuecomment-264851951