Return back to single screen component
This commit is a sort of a rollback of 80e3bf53, returning to a single,
long, ugly component to have an usable prototype by the user that allows
to test it and gather information about bugs, UX/UI changes, etc.

Meanwhile, it will be tested some approaches to split the calculator
into small parts going back to the initial idea to have the ability to
create other calculators screens easily.